+++ lib/src/picker/cupertino_picker_divider.dart
@@ -11,9 +11,7 @@ class CupertinoPickerDivider extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Padding(
-      padding: const EdgeInsets.symmetric(
-        horizontal: 16.0,
-      ),
+      padding: const EdgeInsets.symmetric(horizontal: 16.0),
       child: Divider(
         height: 0.3,
         color: CupertinoColors.separator.resolveFrom(context),